python/twic-position-finder/requirements.txt
python-chess>=1.11.0
requests>=2.28.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic[email]>=2.0.0
boto3>=1.34.0
slowapi>=0.1.9
